Split warning test into new test file, added to travis execution on 2.6 / pypy3
