+++ test_fuzzywuzzy_hypothesis.py
@@ -1,4 +1,3 @@
-import warnings
 from itertools import product
 from functools import partial
 
@@ -76,11 +75,3 @@ def test_identical_strings_extracted(scorer, processor, data):
     assert (choice, 100) in result
 
 
-def test_process_warning():
-    """Check that a string reduced to 0 by processor raises a warning"""
-    query = ':::::::'
-    choices = [':::::::']
-    with warnings.catch_warnings(record=True) as w:
-        result = process.extractOne(query, choices)
-        assert issubclass(w[-1].category, UserWarning)
-        assert result == (query, 0)
